+++ benches/bench-ss-algo-indices.rs
@@ -64,5 +64,12 @@ fn criterion_benchmark(c: &mut Criterion) {
     });
 }
 
-criterion_group!(benches, criterion_benchmark);
+criterion_group! {
+    name = benches;
+    config = Criterion::default()
+        .warm_up_time(std::time::Duration::from_millis(300))
+        .measurement_time(std::time::Duration::from_millis(3000));
+    targets = criterion_benchmark
+}
+//criterion_group!(benches, criterion_benchmark);
 criterion_main!(benches);