+++ benches/create_data.rs
@@ -7,9 +7,13 @@ pub fn create_data() -> (Vec<String>, usize, &'static str, &'static str, &'stati
 }
 
 pub fn create_data_en_1() -> (Vec<String>, usize, &'static str, &'static str, &'static str) {
-    let s1 =
-        "abcdefghijk1234567890".repeat(10) + "ErrWarnAlert" + "abcdefghijklmno".repeat(10).as_str();
-    let s2 = "abcdefghijk1234567890".repeat(10) + "abcdefghijklmno".repeat(10).as_str();
+    let a1 = "You could not possibl"; // 21 bytes
+    let a2 = "my dear Wats"; // 12 bytes
+    let a3 = "This gentleman,"; // 15 bytes
+    let a4 = "Your cases have indee"; // 21 bytes
+    let s1 = a1.repeat(10) + a2 + a3.repeat(10).as_str(); // 372 = 21*10 + 12 + 15*10
+    let s2 = a4.repeat(10) + a3.repeat(10).as_str(); // 360 = 21*10 + 15*10
+
     let mut v: Vec<String> = Vec::new();
     let mut i = 0;
     loop {
@@ -24,18 +28,17 @@ pub fn create_data_en_1() -> (Vec<String>, usize, &'static str, &'static str, &'
         }
     }
     let match_cnt = v.len() / 2;
-    (
-        v,
-        match_cnt,
-        "ErrWarnAlert",
-        "ErrWarnAlert",
-        "*ErrWarnAlert*",
-    )
+    (v, match_cnt, "my dear WatsThis gentleman,", "", "")
 }
 
 pub fn create_data_ja_1() -> (Vec<String>, usize, &'static str, &'static str, &'static str) {
-    let s1 = "吾輩は猫である".repeat(10) + "夏目漱石" + "坊っちゃん".repeat(10).as_str();
-    let s2 = "名前はまだない".repeat(10) + "坊っちゃん".repeat(10).as_str();
+    let a1 = "吾輩は猫である"; // 21 bytes
+    let a2 = "夏目漱石"; // 12 bytes
+    let a3 = "坊っちゃん"; // 15 bytes
+    let a4 = "名前はまだない"; // 21 bytes
+    let s1 = a1.repeat(10) + a2 + a3.repeat(10).as_str(); // 372 = 21*10 + 12 + 15*10
+    let s2 = a4.repeat(10) + a3.repeat(10).as_str(); // 360 = 21*10 + 15*10
+
     let mut v: Vec<String> = Vec::new();
     let mut i = 0;
     loop {
@@ -50,5 +53,5 @@ pub fn create_data_ja_1() -> (Vec<String>, usize, &'static str, &'static str, &'
         }
     }
     let match_cnt = v.len() / 2;
-    (v, match_cnt, "夏目漱石", "夏目漱石", "*夏目漱石*")
+    (v, match_cnt, "夏目漱石坊っちゃん", "", "")
 }