+++ tests/func.rs
@@ -323,3 +323,18 @@ mod func_rsearch_indices {
         assert_eq!(m.next(), None);
     }
 }
+
+#[cfg(test)]
+mod func_str_str_large {
+    use naive_opt::string_search;
+    use naive_opt::string_search_bytes;
+    use naive_opt::string_rsearch;
+    use naive_opt::string_rsearch_bytes;
+    #[test]
+    fn test_large_needle_found() {
+        let haystack = "c11 a 111b12345678901234567890";
+        let needle = "12345678901234567890";
+        search_test!(haystack, needle, Some(10));
+        rsearch_test!(haystack, needle, Some(10));
+    }
+}